+++ src/Products/MeetingCPASLalouviere/adapters.py
@@ -306,26 +306,6 @@ LLO_WAITING_ADVICES_FROM_STATES = {
          # must end with _waiting_advices
          'new_state_id': None,
          },
-        {'from_states': ('proposed_to_president', ),
-         'back_states': ('proposed_to_president', ),
-         'perm_cloned_state': 'validated',
-         'use_custom_icon': False,
-         # default to "validated", this avoid using the backToValidated title that
-         # is translated to "Remove from meeting"
-         'use_custom_back_transition_title_for': ("validated", ),
-         # we can define some back transition id for some back_to_state
-         # if not, a generated transition is used, here we could have for example
-         # 'defined_back_transition_ids': {"validated": "validate"}
-         'defined_back_transition_ids': {},
-         # if () given, a custom transition icon is used for every back transitions
-         'only_use_custom_back_transition_icon_for': ("validated", ),
-         'use_custom_state_title': True,
-         'use_custom_transition_title_for': {},
-         'remove_modify_access': True,
-         'adviser_may_validate': False,
-         # must end with _waiting_advices
-         'new_state_id': None,
-         },
     ),
 }
 adaptations.WAITING_ADVICES_FROM_STATES.update(LLO_WAITING_ADVICES_FROM_STATES)